+++ docs/changes.rst
@@ -9,6 +9,7 @@ Released on ? 2016.
 * Extended parameters for createadmin management command.
 * Extended import_json to be able to handle with existing components.
 * Added support for YAML files.
+* Project owners can now configure translation component details.
 
 weblate 2.8
 -----------